+++ include/BS/RegistryItem.hpp
