+++ include/BS/SwitchingBar.hpp
