+++ include/BS/SwitchingWidget.hpp
