+++ include/BS/WidgetRegistry.hpp
