+++ include/implementation/Global.hpp
