+++ include/implementation/SplitterDecorator.hpp
