+++ include/implementation/SplitterHandle.hpp
