+++ include/implementation/SwitchingCombo.hpp
