+++ src/RegistryItem.cpp
@@ -1,3 +1,3 @@
-#include "include/RegistryItem.hpp"
+#include "include/BS/RegistryItem.hpp"
 
 RegistryItem::RegistryItem(QString name, QWidget* (*widget) (), void (*populateBar) (SwitchingBar*, QWidget*)) : name{name}, widget{widget}, populateBar{populateBar} {}
